Add files via upload
This .html document includes in one file the four components required for the GUI of the webpage: the HTML structure, the CSS styling, as well as the JavaScript scripts and XML required for the server communication. The content of this file is included as a raw literal in the C-compatible Arduino code.